[update] contrib and models
